user state and actions files renamed to fit our schema
