+++ editor/src/app/header/header.component.ts
@@ -5,7 +5,7 @@ import { map } from 'rxjs/operators';
 import { Select, Store } from '@ngxs/store';
 
 import { AppState } from '../app-state/app.state';
-import { UserState } from '../user/user-state';
+import { UserState } from '../user/user.state';
 import { UserStateModel } from '../user/user.state.model';
 
 @Component({
