+++ editor/src/app/sites/sections/tags/section-tags.state.ts
@@ -13,7 +13,7 @@ import {
   AddSectionTagsAction,
   ResetSiteSectionsTagsAction,
   InitSiteSectionsTagsAction} from './section-tags.actions';
-import { UserLoginAction } from '../../../user/user-actions';
+import { UserLoginAction } from '../../../user/user.actions';
 
 
 @State<SectionTagsStateModel>({
